requirements.txt
azure-core
python-qpid-proton
